Smack: Ptrace access check mode
When the ptrace security hooks were split the addition of
a mode parameter was not taken advantage of in the Smack
ptrace access check. This changes the access check from
always looking for read and write access to using the
passed mode. This will make use of /proc much happier.

Targeted for git://git.gitorious.org/smack-next/kernel.git

Change-Id: I59f9f5fb88da2dd08f80a2946afb211db103b16e
Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>