Smack: Make the syslog control configurable
commit 00f84f3f2e9d088f06722f4351d67f5f577abe22
Author: Casey Schaufler <casey@schaufler-ca.com>
Date:   Mon Dec 23 11:07:10 2013 -0800

The syslog control requires that the calling proccess
have the floor ("_") Smack label. Tizen does not run any
processes except for kernel helpers with the floor label.
This changes allows the admin to configure a specific
label for syslog. The default value is the star ("*")
label, effectively removing the restriction. The value
can be set using smackfs/syslog for anyone who wants
a more restrictive behavior.

Change-Id: Iec4c42c4d156fe29915f07a6f40d6c58ad894d6a
Signed-off-by: Casey Schaufler <casey@schaufler-ca.com>