Smack: Assign smack_known_web as default smk_in label for kernel thread's socket
This change fixes the bug associated with sockets owned by kernel threads. These
sockets, created usually by network devices' drivers tasks, received smk_in
label from the task that created them - the "floor" label in the most cases. The
result was that they were not able to receive data packets because of missing
smack rules. The main reason of the access deny is that the socket smk_in label
is placed as the object during smk check, kernel thread's capabilities are
omitted.

Change-Id: Ia4d96cdbde11d9d1facc6da167a475181cafbda3
Signed-off-by: Marcin Lis <m.lis@samsung.com>
Signed-off-by: Rafal Krypa <r.krypa@samsung.com>