smack: check perms for desc. passed via sockets
This is an initial draft of a SMACK patch that will
check all passed descriptors (not only to files) that
are passed via UDS sockets.

This patchset adds a case when the sent descriptor
is created via the socketpair() call, in this case
the connect function is never called, and we need
to label sockets at the time of allocation.

Change-Id: I0decee784c5293dac768f9c753193dc08081921d
Signed-off-by: r.kubiak <r.kubiak@samsung.com>